S15.e.3: extract operator.ts security + cluster fetchers (2483 → 1960 LOC)
Third sub-slice of the S15.e operator.ts decomposition train.

Lifts the remaining five data-fetching functions out of the giant
startDashboard closure into two new fetcher modules grouped by I/O
target:

cli/src/commands/operator/fetchers/security.ts (351 LOC):
  - fetchEgressDomains(sb, kubeContext?)    — was closure-captured kubeContext
  - fetchSecurityState(sb, kubeContext?)    — was closure-captured kubeContext
  - fetchAgtQuick(sb, existing, kubeContext?) — was closure-captured
    securityStates.get(sb.name) + kubeContext

cli/src/commands/operator/fetchers/cluster.ts (188 LOC):
  - fetchMeshHealth(devMode, kubeContext?)    — was closure-captured devMode + kubeContext
  - fetchClusterHealth(devMode, kubeContext?) — was closure-captured devMode + kubeContext

Behavioral change: NONE.
- Bodies byte-identical to originals.
- Mechanical edits: closure-captured kubeContext/devMode/securityStates
  cache lookup are now explicit parameters; call sites in refresh()
  updated to forward them.
- fetchAgtQuick still mutates the passed-in SecurityState in place
  (preserving the existing dashboard cache semantics that keep mesh
  counters alive between full security refreshes).
- Same router endpoints polled (/readyz, /blocklist/status, /agt/status,
  /egress/allowlist, /metrics, /agt/audit, /agt/reputation), same K8s
  checks (NetworkPolicy 'sandbox-policy', secret 'router-admin-token').

Cumulative S15.e delta:
  pre-S15.e: 2894
  e.1:       2739 (-155) — types.ts + helpers.ts
  e.2:       2483 (-256) — fetchers/sandboxes.ts (+ kctl move)
  e.3:       1960 (-523) — fetchers/security.ts + fetchers/cluster.ts
  Cap (800): not yet met; multi-PR sub-train continues with render/, actions/, dialogs/.

Verification:
- operator.ts LOC: 2483 → 1960 (cumulative 2894 → 1960 = -934)
- tsc --noEmit clean / lint 21 warnings (was 27) / build clean
- npm test: 454 pass / 2 skipped pre-existing

Audit: docs/security-audits/2026-04-29-phase2-hotspot-operator-cli-e3.md
