+++ cli/src/commands/operator.ts
@@ -27,6 +27,15 @@ import type {
 } from "./operator/types.js";
 import { timeSince, sumPrometheusCounter, kctl } from "./operator/helpers.js";
 import { fetchSandboxes } from "./operator/fetchers/sandboxes.js";
+import {
+  fetchEgressDomains,
+  fetchSecurityState,
+  fetchAgtQuick,
+} from "./operator/fetchers/security.js";
+import {
+  fetchMeshHealth,
+  fetchClusterHealth,
+} from "./operator/fetchers/cluster.js";
 
 // ── Command ─────────────────────────────────────────────────────────
 
@@ -270,538 +279,6 @@ async function startDashboard(refreshInterval: number, kubeContext?: string, dev
   }
   function stopSpinner() {
     if (spinTimer) { clearInterval(spinTimer); spinTimer = null; }
-  }
-
-  // ── Data fetching ─────────────────────────────────────────────────
-
-
-  async function fetchEgressDomains(sb: SandboxInfo): Promise<EgressDomain[]> {
-    if (!sb.podName) return [];
-    const isDockerAgent = sb.runtime === "docker";
-    const routerCurl = isDockerAgent
-      ? (path: string) => execa("docker", [
-          "exec", sb.podName!,
-          "curl", "-s", "--max-time", "3", `http://localhost:8443${path}`,
-        ], { stdio: "pipe" })
-      : (path: string) => execa("kubectl", kctl([
-          "exec", "-n", sb.namespace, sb.podName!,
-          "-c", "inference-router", "--",
-          "curl", "-s", "--max-time", "3", `http://localhost:8443${path}`,
-        ], kubeContext), { stdio: "pipe", timeout: 10000 });
-
-    try {
-      const [learnedRes, allowRes] = await Promise.allSettled([
-        routerCurl("/egress/learned"),
-        routerCurl("/egress/allowlist"),
-      ]);
-
-      const learnedDomains: Set<string> = new Set();
-      const approvedDomains: Set<string> = new Set();
-
-      if (learnedRes.status === "fulfilled") {
-        const data = JSON.parse((learnedRes.value as any).stdout);
-        for (const d of (data.domains || [])) learnedDomains.add(d);
-      }
-      if (allowRes.status === "fulfilled") {
-        const data = JSON.parse((allowRes.value as any).stdout);
-        for (const d of (data.domains || [])) approvedDomains.add(d);
-      }
-
-      // Merge: domains in allowlist are "approved", rest are "learned" (pending)
-      const results: EgressDomain[] = [];
-      const allDomains = new Set([...learnedDomains, ...approvedDomains]);
-      for (const d of allDomains) {
-        results.push({
-          domain: d,
-          sandbox: sb.name,
-          namespace: sb.namespace,
-          state: approvedDomains.has(d) ? "approved" : "learned",
-        });
-      }
-      // Sort: pending first, then approved
-      results.sort((a, b) => {
-        if (a.state !== b.state) return a.state === "learned" ? -1 : 1;
-        return a.domain.localeCompare(b.domain);
-      });
-      return results;
-    } catch {
-      return [];
-    }
-  }
-
-  /** Poll security-relevant endpoints for a single sandbox. */
-  async function fetchSecurityState(sb: SandboxInfo): Promise<SecurityState> {
-    const state: SecurityState = {
-      sandbox: sb.name,
-      isolation: sb.isolation,
-      runtime: sb.isolation === "confidential" ? "kata-vm" : "runc",
-      seccomp: sb.runtime === "docker" ? "azureclaw-strict"
-               : sb.isolation === "enhanced" ? "azureclaw-strict" : "RuntimeDefault",
-      networkPolicy: false,
-      adminAuth: false,
-      readyz: false,
-      readyzDetail: "unknown",
-      egressMode: "unknown",
-      learnedDomains: 0,
-      allowlistDomains: 0,
-      blocklistDomains: 0,
-      blocklistLearnMode: false,
-      agtEnabled: false,
-      agtAuditEntries: 0,
-      agtAuditIntegrity: false,
-      agtKnownAgents: 0,
-      agtTrustThreshold: 0,
-      agtRecentAudit: [],
-      agtTrustScores: [],
-      agtRelayConnected: false,
-      agtRegistryAgents: 0,
-      agtAmid: "",
-      agtMeshSessions: 0,
-      agtMeshSent: 0,
-      agtMeshReceived: 0,
-      agtTrustUpdates: 0,
-      agtTotalInteractions: 0,
-      agtGovernanceMode: "",
-      agtPolicyEvaluations: 0,
-      agtPolicyDenials: 0,
-      agtPolicyRateLimits: 0,
-      agtEvalLatencyUs: 0,
-      agtBehaviorAlerts: 0,
-      agtBehaviorDetail: [],
-      agtContentFlags: 0,
-      agtPolicyRules: 0,
-      agtReputation: null,
-      agtRelayUrl: "",
-      agtRegistryUrl: "",
-      totalRequests: 0,
-      errorRequests: 0,
-      inputTokens: 0,
-      outputTokens: 0,
-      avgLatencyMs: 0,
-    };
-
-    if (!sb.podName) return state;
-
-    // Use per-sandbox runtime to choose exec path (not global devMode) so the
-    // unified view can query Docker agents via docker-exec and AKS agents via
-    // kubectl-exec in the same operator session.
-    const isDockerAgent = sb.runtime === "docker";
-    const routerExec = isDockerAgent
-      ? (path: string) => execa("docker", [
-          "exec", sb.podName!,
-          "curl", "-s", "--max-time", "3", `http://localhost:8443${path}`,
-        ], { stdio: "pipe" })
-      : (path: string) => execa("kubectl", kctl([
-          "exec", "-n", sb.namespace, sb.podName!,
-          "-c", "inference-router", "--",
-          "curl", "-s", "--max-time", "3", `http://localhost:8443${path}`,
-        ], kubeContext), { stdio: "pipe", timeout: 10000 });
-
-    // Docker agents don't have K8s resources (NetworkPolicy, admin token secret)
-    const k8sCheck = (args: string[]) => isDockerAgent
-      ? Promise.reject("docker-agent")
-      : execa("kubectl", kctl(args, kubeContext), { stdio: "pipe", timeout: 10000 });
-
-    const checks = await Promise.allSettled([
-      // 0: NetworkPolicy (K8s only)
-      k8sCheck(["get", "networkpolicy", "sandbox-policy", "-n", sb.namespace, "-o", "name"]),
-      // 1: Admin token secret (K8s only)
-      k8sCheck(["get", "secret", "router-admin-token", "-n", sb.namespace, "-o", "name"]),
-      // 2: /readyz (body, not just status code)
-      routerExec("/readyz"),
-      // 3: /blocklist/status
-      routerExec("/blocklist/status"),
-      // 4: /agt/status
-      routerExec("/agt/status"),
-      // 5: /egress/allowlist
-      routerExec("/egress/allowlist"),
-      // 6: /metrics (Prometheus text)
-      routerExec("/metrics"),
-      // 7: /agt/audit (last entries)
-      routerExec("/agt/audit"),
-      // 8: /agt/reputation (registry + local trust)
-      routerExec("/agt/reputation"),
-    ]);
-
-    // NetworkPolicy
-    if (checks[0].status === "fulfilled") state.networkPolicy = true;
-
-    // Admin token
-    if (checks[1].status === "fulfilled") state.adminAuth = true;
-
-    // readyz
-    if (checks[2].status === "fulfilled") {
-      const body = ((checks[2].value as any).stdout || "").trim();
-      state.readyz = body.startsWith("ok");
-      state.readyzDetail = body || "ok";
-    }
-
-    // blocklist/status
-    if (checks[3].status === "fulfilled") {
-      try {
-        const bl = JSON.parse((checks[3].value as any).stdout);
-        state.blocklistDomains = bl.domain_count || 0;
-        state.blocklistLearnMode = bl.learn_mode ?? false;
-        state.learnedDomains = bl.learned_domains || 0;
-        state.egressMode = bl.learn_mode ? "learning" : "enforcing";
-      } catch { /* parse fail */ }
-    }
-
-    // agt/status — includes trust_states and inbox count
-    if (checks[4].status === "fulfilled") {
-      try {
-        const agt = JSON.parse((checks[4].value as any).stdout);
-        state.agtEnabled = agt.enabled ?? false;
-        state.agtAuditEntries = agt.audit_entries || 0;
-        state.agtAuditIntegrity = agt.audit_integrity ?? false;
-        state.agtKnownAgents = agt.known_agents || 0;
-        // Trust states from /agt/status response
-        const ts = agt.trust_states || [];
-        state.agtTrustScores = ts.map((a: any) => ({
-          agent: a.agent_id || a.name || "unknown",
-          score: a.score ?? 0,
-          tier: a.tier || (a.score >= 800 ? "Sovereign" : a.score >= 600 ? "Verified" : a.score >= 400 ? "Known" : a.score >= 200 ? "Observed" : "Anonymous"),
-          interactions: a.interactions ?? 0,
-          lastSeen: a.last_interaction || "",
-        }));
-        state.agtRegistryAgents = ts.length;
-        // Mesh metrics from router MeshMetrics counters
-        state.agtMeshSessions = agt.mesh_sessions || 0;
-        state.agtMeshSent = agt.mesh_messages_sent || 0;
-        state.agtMeshReceived = agt.mesh_messages_received || 0;
-        state.agtTrustUpdates = agt.trust_updates || 0;
-        state.agtTotalInteractions = agt.total_interactions || 0;
-        // Native governance stats
-        state.agtGovernanceMode = agt.governance_mode || "";
-        state.agtPolicyEvaluations = agt.policy_evaluations || 0;
-        state.agtPolicyDenials = agt.policy_denials || 0;
-        state.agtPolicyRateLimits = agt.policy_rate_limits || 0;
-        state.agtEvalLatencyUs = agt.eval_latency_avg_us || 0;
-        state.agtBehaviorAlerts = agt.behavior_alerts || 0;
-        state.agtBehaviorDetail = agt.behavior_alerts_detail || [];
-        state.agtContentFlags = agt.content_flags || 0;
-        state.agtPolicyRules = agt.policy_rules || 0;
-        state.agtRelayUrl = agt.relay_url || "";
-        state.agtRegistryUrl = agt.registry_url || "";
-        // If no trust states but governance is enabled, show self
-        if (state.agtEnabled && ts.length === 0) {
-          state.agtTrustScores = [{
-            agent: agt.sandbox || sb.name,
-            score: 500,
-            tier: "Known (self)",
-            interactions: 0,
-            lastSeen: "",
-          }];
-          state.agtRegistryAgents = 1;
-        }
-      } catch { /* parse fail */ }
-    }
-
-    // egress/allowlist
-    if (checks[5].status === "fulfilled") {
-      try {
-        const al = JSON.parse((checks[5].value as any).stdout);
-        state.allowlistDomains = al.count || 0;
-      } catch { /* parse fail */ }
-    }
-
-    // /metrics — parse Prometheus text format
-    if (checks[6].status === "fulfilled") {
-      const metricsText = (checks[6].value as any).stdout || "";
-      state.totalRequests = sumPrometheusCounter(metricsText, "azureclaw_inference_requests_total");
-      state.errorRequests = sumPrometheusCounter(metricsText, "azureclaw_inference_requests_total", { status: "error" });
-      state.inputTokens = sumPrometheusCounter(metricsText, "azureclaw_tokens_total", { direction: "input" });
-      state.outputTokens = sumPrometheusCounter(metricsText, "azureclaw_tokens_total", { direction: "output" });
-
-      // Average latency from histogram sum/count
-      const latSum = sumPrometheusCounter(metricsText, "azureclaw_inference_latency_seconds_sum");
-      const latCount = sumPrometheusCounter(metricsText, "azureclaw_inference_latency_seconds_count");
-      state.avgLatencyMs = latCount > 0 ? Math.round((latSum / latCount) * 1000) : 0;
-    }
-
-    // /agt/audit — extract last few entries
-    if (checks[7].status === "fulfilled") {
-      try {
-        const audit = JSON.parse((checks[7].value as any).stdout);
-        const entries = Array.isArray(audit) ? audit : audit.entries || [];
-        state.agtRecentAudit = entries.slice(-5).map((e: any) => {
-          const action = e.action || e.type || "unknown";
-          const agent = e.agent_id || e.agent || "";
-          const tool = e.tool || "";
-          const result = e.result || e.decision || "";
-          let ts = "";
-          if (e.timestamp) {
-            const d = new Date(e.timestamp);
-            ts = isNaN(d.getTime()) ? "" : d.toLocaleTimeString();
-          }
-          const parts = [action];
-          if (tool) parts.push(`[${tool}]`);
-          if (agent) parts.push(agent.substring(0, 16));
-          if (result) parts.push(`→ ${result}`);
-          return ts ? `${ts} ${parts.join(" ")}` : parts.join(" ");
-        });
-      } catch { /* parse fail */ }
-    }
-
-    // /agt/reputation — registry score + local trust
-    if (checks[8].status === "fulfilled") {
-      try {
-        const rep = JSON.parse((checks[8].value as any).stdout);
-        // AMID (cryptographic identity from registry lookup)
-        if (rep.amid) state.agtAmid = rep.amid;
-        // Registry reputation (from agentmesh-registry Postgres)
-        if (rep.registry) {
-          const r = rep.registry;
-          state.agtReputation = {
-            score: r.score ?? 0,
-            tier: r.tier || "unknown",
-            completionRate: r.completion_rate ?? 0,
-            totalSessions: r.total_sessions ?? 0,
-            feedbackCount: r.feedback_count ?? 0,
-            avgFeedback: r.average_feedback ?? 0,
-            tags: (r.tags || []).map((t: any) => ({ tag: t.tag || "", count: t.count || 0 })),
-          };
-        }
-        // Local trust store (from router in-memory)
-        const local = rep.local_trust || [];
-        if (local.length > 0) {
-          state.agtTrustScores = local.map((a: any) => ({
-            agent: a.agent_id || a.name || "unknown",
-            score: a.score ?? 0,
-            tier: a.tier || (a.score >= 800 ? "Sovereign" : a.score >= 600 ? "Verified" : a.score >= 400 ? "Known" : "Anonymous"),
-            interactions: a.interactions ?? 0,
-            lastSeen: a.last_interaction || "",
-          }));
-          state.agtRegistryAgents = local.length;
-        }
-      } catch { /* parse fail */ }
-    }
-
-    return state;
-  }
-
-  // ── Fast AGT Poll (every cycle) ─────────────────────────────────
-  // Lightweight: only fetches /agt/status per sandbox to keep trust scores
-  // and mesh counters alive between full security refreshes.
-
-  async function fetchAgtQuick(sb: SandboxInfo): Promise<void> {
-    if (!sb.podName) return;
-    const existing = securityStates.get(sb.name);
-    if (!existing?.agtEnabled) return;
-
-    try {
-      const isDockerAgent = sb.runtime === "docker";
-      const { stdout } = isDockerAgent
-        ? await execa("docker", [
-            "exec", sb.podName,
-            "curl", "-s", "--max-time", "2", "http://localhost:8443/agt/status",
-          ], { stdio: "pipe" })
-        : await execa("kubectl", kctl([
-            "exec", "-n", sb.namespace, sb.podName,
-            "-c", "inference-router", "--",
-            "curl", "-s", "--max-time", "2", "http://localhost:8443/agt/status",
-          ], kubeContext), { stdio: "pipe", timeout: 8000 });
-
-      const agt = JSON.parse(stdout);
-      const ts = agt.trust_states || [];
-      if (ts.length > 0) {
-        existing.agtTrustScores = ts.map((a: any) => ({
-          agent: a.agent_id || a.name || "unknown",
-          score: a.score ?? 0,
-          tier: a.tier || (a.score >= 800 ? "Sovereign" : a.score >= 600 ? "Verified" : a.score >= 400 ? "Known" : a.score >= 200 ? "Observed" : "Anonymous"),
-          interactions: a.interactions ?? 0,
-          lastSeen: a.last_interaction || "",
-        }));
-        existing.agtRegistryAgents = ts.length;
-      }
-      existing.agtMeshSessions = agt.mesh_sessions || 0;
-      existing.agtMeshSent = agt.mesh_messages_sent || 0;
-      existing.agtMeshReceived = agt.mesh_messages_received || 0;
-      existing.agtTrustUpdates = agt.trust_updates || 0;
-      existing.agtTotalInteractions = agt.total_interactions || 0;
-      existing.agtAuditEntries = agt.audit_entries || 0;
-      existing.agtAuditIntegrity = agt.audit_integrity ?? false;
-      existing.agtGovernanceMode = agt.governance_mode || existing.agtGovernanceMode;
-      existing.agtPolicyEvaluations = agt.policy_evaluations || 0;
-      existing.agtPolicyDenials = agt.policy_denials || 0;
-      existing.agtPolicyRateLimits = agt.policy_rate_limits || 0;
-      existing.agtEvalLatencyUs = agt.eval_latency_avg_us || 0;
-      existing.agtBehaviorAlerts = agt.behavior_alerts || 0;
-      existing.agtBehaviorDetail = agt.behavior_alerts_detail || [];
-      existing.agtContentFlags = agt.content_flags || 0;
-      existing.agtPolicyRules = agt.policy_rules || existing.agtPolicyRules;
-    } catch { /* non-fatal — full refresh on next TIER_DETAIL cycle */ }
-  }
-
-  // ── Mesh Health ──────────────────────────────────────────────────
-
-  async function fetchMeshHealth(): Promise<MeshHealth> {
-    const result: MeshHealth = { relayReady: false, registryReady: false, registryPods: 0, registryReadyPods: 0 };
-    if (devMode) {
-      try {
-        const { stdout } = await execa("docker", ["ps", "--filter", "name=relay", "--filter", "status=running", "--format", "{{.Names}}"], { stdio: "pipe", timeout: 5000 });
-        result.relayReady = stdout.trim().length > 0;
-      } catch {}
-      try {
-        const { stdout } = await execa("docker", ["ps", "--filter", "name=registry", "--filter", "status=running", "--format", "{{.Names}}"], { stdio: "pipe", timeout: 5000 });
-        result.registryReady = stdout.trim().length > 0;
-        result.registryPods = result.registryReady ? 1 : 0;
-        result.registryReadyPods = result.registryPods;
-      } catch {}
-      return result;
-    }
-    try {
-      const { stdout } = await execa("kubectl", kctl([
-        "get", "pods", "-n", "agentmesh", "-o",
-        `jsonpath={range .items[*]}{.metadata.labels.app}{"||"}{.status.conditions[?(@.type=="Ready")].status}{"\\n"}{end}`,
-      ], kubeContext), { stdio: "pipe", timeout: 10000 });
-      for (const line of stdout.trim().split("\n")) {
-        if (!line) continue;
-        const [app, ready] = line.split("||");
-        if (app?.includes("relay") && ready === "True") result.relayReady = true;
-        if (app?.includes("registry")) {
-          result.registryPods++;
-          if (ready === "True") { result.registryReady = true; result.registryReadyPods++; }
-        }
-      }
-    } catch {}
-    return result;
-  }
-
-  // ── Cluster Health ────────────────────────────────────────────────
-
-  async function fetchClusterHealth(): Promise<ClusterHealth> {
-    const result: ClusterHealth = {
-      apiLatencyMs: -1, apiReachable: false,
-      nodes: [], quotas: [], pvcs: [], warnings: [],
-    };
-
-    if (devMode) {
-      result.apiReachable = true;
-      result.apiLatencyMs = 0;
-      try {
-        const { stdout } = await execa("docker", ["info", "--format", "{{.NCPU}}|{{.MemTotal}}|{{.ServerVersion}}"], { stdio: "pipe" });
-        const [cpus, mem, version] = stdout.trim().split("|");
-        const memGb = (parseInt(mem || "0", 10) / 1073741824).toFixed(1);
-        result.nodes = [{
-          name: "docker-host",
-          pool: "local",
-          status: "Ready",
-          version: version || "",
-          cpuCores: `${cpus}`,
-          cpuPct: "-",
-          memBytes: `${memGb}Gi`,
-          memPct: "-",
-          os: "Docker",
-          runtime: `docker://${version || ""}`,
-        }];
-      } catch { /* docker info fail */ }
-      return result;
-    }
-
-    // All queries in parallel
-    const [apiRes, nodesRes, topRes, quotaRes, pvcRes, eventsRes] = await Promise.allSettled([
-      // 0: API server latency
-      (async () => {
-        const start = Date.now();
-        await execa("kubectl", kctl(["get", "ns", "--no-headers"], kubeContext),
-          { stdio: "pipe", timeout: 10000 });
-        return Date.now() - start;
-      })(),
-      // 1: Node info
-      execa("kubectl", kctl([
-        "get", "nodes", "-o",
-        `jsonpath={range .items[*]}{.metadata.name}|{.metadata.labels.agentpool}|` +
-        `{.status.conditions[?(@.type=="Ready")].status}|{.status.nodeInfo.kubeletVersion}|` +
-        `{.status.nodeInfo.osImage}|{.status.nodeInfo.containerRuntimeVersion}{\"\\n\"}{end}`,
-      ], kubeContext), { stdio: "pipe" }),
-      // 2: kubectl top nodes
-      execa("kubectl", kctl(["top", "nodes", "--no-headers"], kubeContext),
-        { stdio: "pipe", timeout: 10000 }),
-      // 3: Resource quotas
-      execa("kubectl", kctl([
-        "get", "resourcequotas", "-A", "-o",
-        `jsonpath={range .items[*]}{.metadata.namespace}|{.status.used.cpu}|{.status.hard.cpu}|` +
-        `{.status.used.memory}|{.status.hard.memory}{\"\\n\"}{end}`,
-      ], kubeContext), { stdio: "pipe" }),
-      // 4: PVCs
-      execa("kubectl", kctl([
-        "get", "pvc", "-A", "-o",
-        `jsonpath={range .items[*]}{.metadata.namespace}|{.metadata.name}|{.status.phase}|` +
-        `{.spec.resources.requests.storage}{\"\\n\"}{end}`,
-      ], kubeContext), { stdio: "pipe" }),
-      // 5: Warning events
-      execa("kubectl", kctl([
-        "get", "events", "-A", "--field-selector", "type=Warning",
-        "--sort-by=.lastTimestamp", "-o",
-        `jsonpath={range .items[-8:]}{.lastTimestamp}|{.reason}|` +
-        `{.involvedObject.kind}/{.involvedObject.name}|{.message}{\"\\n\"}{end}`,
-      ], kubeContext), { stdio: "pipe" }),
-    ]);
-
-    // Parse API latency
-    if (apiRes.status === "fulfilled") {
-      result.apiLatencyMs = apiRes.value as number;
-      result.apiReachable = true;
-    }
-
-    // Parse nodes
-    if (nodesRes.status === "fulfilled") {
-      const topMap = new Map<string, { cpu: string; cpuPct: string; mem: string; memPct: string }>();
-      if (topRes.status === "fulfilled") {
-        for (const line of (topRes.value as any).stdout.trim().split("\n").filter(Boolean)) {
-          const parts = line.trim().split(/\s+/);
-          if (parts.length >= 5) {
-            topMap.set(parts[0], { cpu: parts[1], cpuPct: parts[2], mem: parts[3], memPct: parts[4] });
-          }
-        }
-      }
-
-      for (const line of (nodesRes.value as any).stdout.trim().split("\n").filter(Boolean)) {
-        const [name, pool, ready, version, os, runtime] = line.split("|");
-        if (!name) continue;
-        const top = topMap.get(name);
-        result.nodes.push({
-          name, pool: pool || "-",
-          status: ready === "True" ? "Ready" : "NotReady",
-          version: version || "-",
-          cpuCores: top?.cpu || "-", cpuPct: top?.cpuPct || "-",
-          memBytes: top?.mem || "-", memPct: top?.memPct || "-",
-          os: os || "-", runtime: runtime || "-",
-        });
-      }
-    }
-
-    // Parse quotas
-    if (quotaRes.status === "fulfilled") {
-      for (const line of (quotaRes.value as any).stdout.trim().split("\n").filter(Boolean)) {
-        const [ns, cpuUsed, cpuHard, memUsed, memHard] = line.split("|");
-        if (ns) result.quotas.push({ namespace: ns, cpuUsed: cpuUsed || "0", cpuHard: cpuHard || "-", memUsed: memUsed || "0", memHard: memHard || "-" });
-      }
-    }
-
-    // Parse PVCs
-    if (pvcRes.status === "fulfilled") {
-      for (const line of (pvcRes.value as any).stdout.trim().split("\n").filter(Boolean)) {
-        const [ns, name, phase, size] = line.split("|");
-        if (ns) result.pvcs.push({ namespace: ns, name: name || "-", phase: phase || "Unknown", size: size || "-" });
-      }
-    }
-
-    // Parse warnings
-    if (eventsRes.status === "fulfilled") {
-      for (const line of (eventsRes.value as any).stdout.trim().split("\n").filter(Boolean)) {
-        const [time, reason, object, ...rest] = line.split("|");
-        if (reason) result.warnings.push({
-          time: time ? timeSince(new Date(time)) : "-",
-          reason,
-          object: (object || "-").substring(0, 40),
-          message: (rest.join("|") || "").substring(0, 60),
-        });
-      }
-    }
-
-    return result;
   }
 
   // ── Actions ───────────────────────────────────────────────────────
@@ -1679,14 +1156,14 @@ async function startDashboard(refreshInterval: number, kubeContext?: string, dev
 
       if (fetchDetail) {
         promises.push(
-          Promise.allSettled(running.map((s) => fetchEgressDomains(s))).then((settled) => {
+          Promise.allSettled(running.map((s) => fetchEgressDomains(s, kubeContext))).then((settled) => {
             egressByAgent = new Map();
             for (let i = 0; i < running.length; i++) {
               const r = settled[i];
               if (r.status === "fulfilled") egressByAgent.set(running[i].name, r.value);
             }
           }),
-          Promise.allSettled(running.map((s) => fetchSecurityState(s))).then((settled) => {
+          Promise.allSettled(running.map((s) => fetchSecurityState(s, kubeContext))).then((settled) => {
             securityStates = new Map();
             for (let i = 0; i < running.length; i++) {
               const r = settled[i];
@@ -1697,14 +1174,14 @@ async function startDashboard(refreshInterval: number, kubeContext?: string, dev
       } else {
         // Fast AGT-only poll on non-detail cycles to keep mesh data alive
         promises.push(
-          Promise.allSettled(running.map((s) => fetchAgtQuick(s))),
+          Promise.allSettled(running.map((s) => fetchAgtQuick(s, securityStates.get(s.name), kubeContext))),
         );
       }
 
       if (fetchCluster) {
         promises.push(
-          fetchClusterHealth().then((d) => { clusterData = d; }).catch(() => {}),
-          fetchMeshHealth().then((d) => { meshHealth = d; }).catch(() => {}),
+          fetchClusterHealth(devMode, kubeContext).then((d) => { clusterData = d; }).catch(() => {}),
+          fetchMeshHealth(devMode, kubeContext).then((d) => { meshHealth = d; }).catch(() => {}),
         );
       }
